webapp/requirements.txt
pandas
paho-mqtt==1.6.1
streamlit
streamlit-lottie
pydeck